# Web Portal: Optimize speed of `npm start`