--- a/webportal/.gitignore
+++ b/webportal/.gitignore
@@ -1,7 +1,6 @@
 *.zip
 job-docs.md
 job-submit.example.json
-src/app/config/
 
 # gitignore file for node js
 # Created by GitHub (https://github.com/github/gitignore/blob/master/Node.gitignore)