--- a/webportal/config/preprocess.js
+++ b/webportal/config/preprocess.js
@@ -19,8 +19,6 @@
 // module dependencies
 const fse = require('fs-extra');
 const helpers = require('./helpers');
-const webportalConfig = require('./webportal.config');
-
 
 // copy docs to app
 fse.copySync(
@@ -30,8 +28,3 @@ fse.copySync(
 fse.copySync(
     helpers.root('../examples/tensorflow/tensorflow.distributed-cifar10.json'),
     helpers.root('src/app/job/job-submit/job-submit.example.json'));
-
-// write env config
-fse.outputJsonSync(
-    helpers.root('src/app/config/webportal.config.json'),
-    webportalConfig);