--- a/webportal/package.json
+++ b/webportal/package.json
@@ -74,7 +74,7 @@
     "prebuild": "npm run clean && node config/preprocess.js",
     "build": "webpack --config webpack.config.js --progress --profile",
     "build:dev": "npm run build -- --debug --devtool eval-cheap-module-source-map --output-pathinfo --watch",
-    "prestart": "envsub --env-file .env --system src/app/env.js dist/scripts/env.js && ping -c 1 -W 1 172.17.0.1",
+    "prestart": "envsub --env-file .env --system src/app/env.js dist/scripts/env.js",
     "start": "node server",
     "yarn": "yarn --ignore-engines"
   }