--- a/webportal/src/app/config/webportal.config.js
+++ b/webportal/src/app/config/webportal.config.js
@@ -1,0 +1,1 @@
+module.exports = window.ENV