--- a/webportal/src/app/env.js
+++ b/webportal/src/app/env.js
@@ -1,0 +1,10 @@
+/* eslint-disable */
+window.ENV = {
+  restServerUri: '${REST_SERVER_URI}',
+  prometheusUri: '${PROMETHEUS_URI}',
+  yarnWebPortalUri: '${YARN_WEB_PORTAL_URI}',
+  grafanaUri: '${GRAFANA_URI}',
+  k8sDashboardUri: '${K8S_DASHBOARD_URI}',
+  k8sApiServerUri: '${K8S_API_SERVER_URI}',
+  exporterPort: '${EXPORTER_PORT}'
+};