--- a/webportal/src/app/job/job-view/job-view.component.js
+++ b/webportal/src/app/job/job-view/job-view.component.js
@@ -34,7 +34,7 @@ const jobDetailTableComponent = require('./job-detail-table.component.ejs');
 const jobDetailConfigInfoModalComponent = require('./job-detail-config-info-modal.component.ejs');
 const jobDetailSshInfoModalComponent = require('./job-detail-ssh-info-modal.component.ejs');
 const loading = require('../loading/loading.component');
-const webportalConfig = require('../../config/webportal.config.json');
+const webportalConfig = require('../../config/webportal.config.js');
 const userAuth = require('../../user/user-auth/user-auth.component');
 
 let table = null;