--- a/webportal/src/app/layout/layout.component.ejs
+++ b/webportal/src/app/layout/layout.component.ejs
@@ -114,6 +114,6 @@
     <div class="control-sidebar-bg"></div>
 
   </div>
-
+  <script src="scripts/env.js"></script>
 </body>
 </html>