… webportal/config/webportal.config.js