1. Move `npm run build` run in image build stage, decouple with env.
2. Inject env to `window.ENV` in `env.js`, refernced by all pages.
3. Rename `webportal.config.json` to js, which exports `window.ENV`.

Closes #936