@Gerhut 2018-08-01
Docs will be updated after #1000 merged