@Gerhut 2018-08-02
Benchmarks:


Optimization \ Stage | Build | Run
-|-|-
without |56.738s|30.9s
with | 1:32.97s | 0.3s


Build:
- `time docker build --tag webportal .`
- Record total seconds

Run:
- `docker run --detach --name container_name image_tag`
- `docker logs --timestamp container_name`
- Record timestamp of first line & listening line, calculate difference

Without optimization:
- Build: **56.738s**
- Run: **8:19:18.7 - 8:18:47.8 = 30.9s**